test to verify reuse of existing working repo
Test to verify that it is reused when valid and not when invalid.
Currently the invalid one fails, as there appears to be some locking
between the initial clone during the test and the deletion during the
gradle run. I opened an issue [1] to address a potential bug in Grgit
that may be causing this.

[1] https://github.com/ajoberstar/grgit/issues/132