Support all valid Git URI-ish
Remotes can use a variety of URI-like values. The checks to validate
that an existing repo is pointing to the correct repository need to
normalize the uris to ensure that minor differences in syntax are
ignored if they are effectively equal. Previously, URL was used for this
normalization, but that cut out the SSH and similar uri-ish types
supported by Git. Using the URIish class from JGit fixes this problem.