Always close the repo
Not closing the repo seemed to hold a lock on the repo directory when
you use the Gradle daemon. I can't actually test that with TestKit since
that doesn't seem to present the error, presumably because the daemons
are shut down after each test. Either way closing the repo seems like
good hygiene.

This is a fix for #16.