Quit burying exceptions during reset task
The reset task was burying exceptions as a way to indicate that the
branch being targeted didn't exist. This would indicate a need to
trigger creation of an orphaned branch.

Burying the exceptions causes actual issues with someone configuring
authentication or the remote URI incorrectly. Can be improved by testing
for the existence of the branch before trying to fetch it.

This resolves #25 by reporting the