Prevent unknown property error during close task
The close task is set up as a finalizer, meaning it runs even when the
tasks preceding it have failed. In some cases, this means the repo
property that it's trying to close may not exist yet. The task now
guards against this by checking if the property exists before trying to
close the repo.