Don't use input/outputs for tasks
As of Gradle 4.2 [1], when execution of a task begins, any "stale" outputs
will be deleted. All files under build and any declared outputs are
considered managed by Gradle. If other files happen to be in there,
Gradle will delete them.

The reset task did not declare any outputs, but created the
build/gitPublish directory and cloned the repo in there. When the copy
task rolls around, which is a built in task with declared
inputs/outputs, it sees some that cloned repo as stale files in its
output directory, since the copy task didn't put them there. Since they
are also under build, Gradle considers itself the owner and thinks it's
safe to delete them.

Since the inputs/outputs of the tasks in this plugin both overlap and
sometimes aren't clear files in / files out situations, it is easier for
now to drop using them altogether.

This commit drops the use of the Copy task, in favor of a ad-hoc task
that uses project.copy. This was the only task that had any
inputs/outputs declared before, so this issue should no longer be able
to present.

This fixes #35.

[1] https://docs.gradle.org/4.2/release-notes.html#safer-handling-of-stale-output-files