Remove memoization of Grgit
When switching to providers, there was the issue that each task that
called the provider would potentially open a new Grgit instance, which
weren't going to be closed properly.

With 1.0.0, we switched to use a Guava memoized Supplier, which was
intended to deal with that issue. It's not entirely clear if it is
causing the issue, but a couple people have noticed getting a
NoRemoteRepositoryException: origin: not found.

This change leverages a different aspect of Providers where I can have
the Reset task be the one to open the Grgit instance, and just set the
opened instance in the Property, which I can then pass to the other
tasks. That way the Property's value isn't based on a Provider that will
be re-evaluated for each task. Even if this doesn't address the issue,
it's probably a decent improvement as it drops the Guava dependency.

This will hopefully address #51.