Don't close grgit if it's not present
The previous change to leverage the reset task to set the grgit property
meant that if reset doesn't run, the grgit property won't be set. Gradle
doesn't let you access an unset property, so you get a failure like
this:

    * What went wrong:
    No value has been specified for this provider.

So we need to just do a safety check before trying to close grgit.

This fixes #51, since a user confirmed the previous change resolved
the origin not found scenario.