Initialize Grgit based on Gradle version
If running on a configuration cache compatible Gradle version, Grgit
will be initialized from a BuildService that automatically closes Grgit
at the end of the build. That way the buil lifecycle listener
registration can be avoided.
If running on an earlier Gradle version the old behavior will be used.

Fixes #93