chore: Make signing tests more robust
Old tests didn't verify the signatures were produced and also assumed
that signing config wasn't set up wherever you ran tests. Since I do
use signed commits now for my own work, this was a bad assumption and
made the tests fail locally.

Now we check for either the signature to be present or that signing
failed because no secret key was found.

For disabling signing, we ensure that no signature was added to the commit.