patch: Use reference only as an alternate obj store
Old behavior was to use the reference repo (usually the project's
checkout repo) as a full remote that we would fetch from. However, Git
has a more direct way to reference an existing repo, using
`objects/info/alternates`. This mechanism will just let you reuse
objects that already exist in the reference repo rather than fetching
from the real remotes.

So now only origin is a remote and we just try to reuse objects from
the remote. This also simplifies things where we had to check if the
branch we cared about existed. Now, we don't care and if the objects
are relevant they get used and if not, it should be no harm, no foul.

This will hopefully address the issue that reported in #112.